# perf(ext/node): cache `IncomingMessageForServer.headers`